Export keep_tip
