+++ inst/include/TreeTools.h
@@ -3,6 +3,7 @@
 
 #include "TreeTools/types.h"
 #include "TreeTools/root_tree.h"
+#include "TreeTools/keep_tip.h"
 
 
 #endif // TreeTools_H_GEN_
